requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
flask-jwt-extended>=4.6.0
flask-marshmallow>=1.2.0
marshmallow-sqlalchemy>=1.0.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
gunicorn>=21.2.0
flask-wtf>=1.2.1
email-validator>=2.1.0.post1
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-flask>=1.3.0
flask-login>=0.6.3
